feat: Simple Mechanism for Settings Storage
This commit actually includes quite a few things.

I had originally intended to implement the first 'play' command, but
this then turned into needing a Source Port, then a mechanism for
storing the Source Ports that had been configured.

There's an implementation of a SourcePort struct and a constructor
method for that, and then a partial implementation of a command for
adding a Source Port. That then led to where I would actually store
the Source Port information, hence the settings storage.

I've decided just to store everything in a simple file-based database.
This will be the case for the settings for the application and for any
user settings. I don't feel like I need anything more, at least at this
stage anyway.

There's a little repository struct that will save and retrieve the
settings, but it won't validate them. I decided to just have the
command handlers do the validation. The repository has been defined
using a trait, so that I can provide mocked ones for testing.