feat: Add Source Port Command
Implements the `tdl source-port add` command.

The command uses the repository to retrieve the current settings, then
adds the new source port to them, and saves them again. There's a check
to make sure a duplicate isn't added.

This also includes using a hard-coded path when the settings file is
instantiated, but this will be swapped out soon when some integration
tests are written.