refactor: Source Port to Type/Version Pair
Rather than allowing the user to provide a 'name' for the Source Port,
the types of possible Source Ports are provided by an Enum. The version
field is retained, and the Source Port is determined to be unique by
combining the type and the version.

The Profile was updated to specify a type and version reference to the
source port rather than just a string.

All the tests are passing.