feat: Initial Play Command
The initial `play` command isn't actually a completely working
implementation, but this makes good progress. What's missing is
the setting of the working directory for the launched process and the
lookup of the specfied Megawad.

What this does provide is a set of integration tests that use a fake
source port binary that will print out the arguments it was passed.
This can then be used for the integration tests to assert that the
correct arguments are being passed to any given source port, which is
the only thing this code has control over or really cares about anyway.

So far it looks up the selected profile, using the default if one isn't
supplied, and then gets the source port from the profile.