feat: WAD LsDir Command
This command parses the header and the 'directory' for the WAD at the
given path. After that, it simply just lists all the directory entries.
Each directory entry has the location and size of the 'lumps' in the
WAD, which is the actual data for the maps and whatnot.

This command isn't very useful, but it was a good exercise to write the
code for parsing the WAD. Although, since we're not really parsing the
actual data contained in the WAD, I've actually called the main struct
'WadMetadata', since to me, that's what the directory really is.

There are integration tests to check the IWADs are parsed correctly,
for both DOOM.WAD and DOOM2.WAD. As per the comment on the test file,
these will need to be confined to run on a private machine where the
WADs have been legally obtained.

The next step now will be to use this parsing to import the actual
IWADs into the collection.