feat: IWAD Import Command
Defines WadEntry and MapInfo structs, which will be used to save
imported WADs (either IWAD or PWAD) and present that information to the
user.

A new 'storage' module is added that contains an ObjectRepository
struct. This is for saving objects like an imported WadEntry to a file,
using JSON as the format. It uses the word "Object" because it's
intended to be generic, and you could use this to persist any other
structs you need to save. I can think of at least one at the moment,
which would be the 'denormalized' entries for doing fuzzy finding.

The FileSettingsRepository will be moved into this module (and also
renamed) as part of another refactor commit.

A new struct is also defined for storing user settings. The way the
user settings will work is, you can either define one variable, which
is `TDL_DOOM_HOME_PATH`, or you will be able to define each of the
configurable elements individually (either using environment variables
or from reading a settings file). Right now, we've just implemented the
first of those.

Functions have been provided for retrieving the path of the directory
where the app settings are stored and for retrieving the user settings.
We offer the user the ability to set the app settings directory by
using the `TDL_SETTINGS_PATH` variable. Otherwise, the default location
will be `~/.config/tdl` on Linux. Windows support will be added later
and will probably use the `%APPDATA&` variable as the default. For user
settings, right now, the mechanism provided is to set a
`TDL_DOOM_HOME_PATH` variable and have all the configurable paths set
relative to that location. More options will be provided later.

The existing 'SettingsRegistry' struct will be renamed to
'AppSettings', but I'll do that as a separate refactor commit.

There's integration tests defined to make sure each IWAD is imported
with the correct information. In the case of IWADs, the information
is being stored directly in this binary, since things like the map
titles weren't included as text metadata in the WAD, but instead were
only defined as graphical text objects. It wasn't until much later that
things like MAPINFO lumps were introduced; these will be parsed when we
get to the WAD importing commands.

These integration tests are going to make the assumption that the IWAD
entries stored by the application are being saved and stored correctly.
I feel this is justified because the ObjectRepository has its own set of
tests covering that functionality.