feat: Fuzzy Find for Selecting Map
Implements a simple fuzzy find using the `skim` library.

Retrieve all the map info entries and then pipe them through the fuzzy
finder provided by `skim`. If this gets too slow when lots of WADs are
added, it can be made concurrent to make it faster. Will need to do some
benchmarks later.

This code doesn't have automated tests because it necessarily involves
user input. I tested it manually and it works OK.

For now, the code isn't being called, but this will be changed shortly.