feat: Select IWAD for play command
The IWAD has now been turned into an optional argument, and if it's
not specified, the fuzzy finder selection will run.

We use the ObjectRepository to look up the specified WadEntry for the
IWAD. This isn't strictly necessary for IWADs, but it will become
necessary for WADs, as some WADs use different cases for the file names.
The file name the WAD was imported with is going to be stored in the
WadEntry.