feat: Select Map for play Command
A `--map` argument is added to the `play` command to allow the user to
select a map to play. The argument accepts the DOOM or DOOM2 map number
formats, namely "MAP04" or "E1M4".

This is then converted into the `-warp` argument expected by the source
port.

An extra field was added to the MapInfo struct to specify the number of
the map to warp to.

If the user doesn't specify the `--megawad` argument, the fuzzy finder
is launched to select the WAD and map, and that was working correctly.

This is quite a good milestone for perhaps a 0.1.0 release. I think I
could now setup CI and test how it works on Windows.