feat: Select Map for play Command on Windows
It turned out that the skim library sadly doesn't work on Windows, and
it's not clear when it will be supported. I had commented on the open
issue on Github to say I would be willing to help with support for
Windows, but currently I just have no idea how I would even start to go
about doing it.

The Windows implementation is to pipe the search string out to the
external `fzf` binary. This does introduce another dependency, but until
skim has support for Windows, this will need to do.

As per the Linux implementation, this doesn't implement any preview for
the search items.