refactor: move and refactor settings storage
The SettingsRegistry and FileSettingsRepository were renamed to AppSettings and
AppSettingsRepository, respectively.

The AppSettingsRepository was then moved to the storage module, to sit alongside the
ObjectRepository.

From this point on I'm going to be using a new format for my commit messages. Subject lines will be
permitted to be 60 characters long and the line length of the body will be permitted to be 100
characters long. The subject will also be all lower case. I'll add the commit linter to CI shortly.