refactor: lsdir tests to use downloaded lswad
These tests were setup to look for lswad at a particular hard coded location. This has now been
changed to download the xwadtools archive from my own mirrored version of it, and extract that to a
temporary location. The test will then run lswad from there.

This could have been done as something like a pre-test Makefile step, but implementing it in Rust
in the tests means it's a cross platform solution.

There was another cross platform change made to the play tests to get those to run on Windows.