ci: workflow for releasing tdl
A workflow for releases:
* Build the code for Linux, Windows and macOS.
* Put the executables in tar archives.
* Generate a description for the release that includes the change log.
* Create a release with the generated description.
* Upload the tar archives to the release.

A file has been added that maintains the version number and a change log has also been added. The
process for doing a release will be to create a commit with a `release:` type and change the version
and update the change log as part of that commit.

There's a Bash script that generates the release description, but the collection of the change log
is done using Python because that would have been so much more hassle using Bash.