feat: profile add command in interactive mode
If the user runs the `profile add` command without specifying the `--name` argument, the profile
will be added using an interactive mode, which opens a template profile document in an editor,
giving the user the ability to specify the options for the profile. There are lots of options, so
this is to make it less cumbersome.

The template document uses the 'hjson' format, which is for human readable json. This format allows
comments in the document, which is good for describing what each of the settings controls and what
valid values might be for them. There was a new crate added for this.

The editor to use is specified by the `EDITOR` variable. This is required to be set, rather than
falling back to any defaults.

Once the editor is closed, the temporary file that was saved is parsed by the hjson library and the
profile is added.