feat: profile ls command
A simple command that just lists all the profiles that have been added. They are printed out using
the `prettytable` crate.

A test helper library was created for common functions. This only has a single function at the
moment but I have no doubt that it will expand over time.

There's now been an additional job added to the PR workflow to run integration tests on all 3
platforms. Right now it's only running these profile tests because those are not dependent on having
the IWADs present.

Also fixed a couple of clippy warnings. I thought the CI would be picking these up and would have
done so when the last PR was merged. Will need to investigate this.