refactor: redefine source ports list
Today I put together a list of source ports to support:
https://gist.github.com/jacderida/dacff1af5be894233624a9ff71b8fefd

Eventually, this will probably appear in the README, but what's relevant for this commit is that the
source port type enum was updated to reflect this list.

The `SourcePort` struct was renamed to `InstalledSourcePort` as there's going to need to be a
distinction between the source ports that are available and the source ports that have actually been
installed, since a command like `source-port ls` are actually going to list the supported source
ports and their latest versions, as opposed to the ports that have actually been downloaded and
installed.