refactor: unit test name format
The new format for unit tests is as follows:
<module>::<struct>::<function>::should_X

The rest of the test name after the "should" tries to succinctly describe the scenario for the test,
based on what the outcome of it is going to be. I tried using much more verbose "given, then" test
names that described the scenario in much more detail, but found them to be a bit unwieldy in the
test run. The tests themselves can be inspected for more detail.

When it's not a struct that's being tested but a function, the format is:
<module>::<function>::should_X

There's a bit of a special case for the commands module, which is:
commands::<submodule>::<subcommand>::should_X

For example: `commands::profile::add::should_save_the_first_profile`.