refactor: rename source port type enum
Rename SourcePortType enum to SourcePort, because in doing the work for listing the available source
ports, I realised that it's not really a 'type' as such, it's just referring to the source port
itself.

Command arguments were also subsequently renamed for this change and I think it actually makes the
commands easier to understand.

This will support further changes for having different InstalledSourcePort and AvailableSourcePort
struct types, which will both have a `name` member that will be of `SourcePort` type.