feat: object repository can delete entries
Deletes a file that was previously saved by the repository.

The new feature for listing the source ports available to install is going to use the Github's API
to query for releases that are available. As it turns out though, Github has quite stringent rate
limits, so the creation of a cache was necessary. This new function on the object repository was
written to support deletion of stale cache items. It would have been possible for the `save`
function to be modified to just overwrite any existing entries, but I just decided against this, in
case I came into any complications with WAD entries being saved over.

This method is idempotent. It checks for the existence of the file to delete so that it doesn't
return an error if you try to delete something that doesn't exist.