feat: add release cache path to app settings
Another piece of functionality to support operations related to source port releases. There's going
to be a cache for the release info because Github's API has pretty stringent usage limits. The app
settings are extended to describe where the files for the cache will be saved.

This also changes the behaviour of the app settings repository to serialize default settings if the
settings file doesn't already exist, as opposed to writing out the string based representation of
the empty app settings struct. Quite honestly, I'm not sure why I didn't do this in the first place.
While the default settings are being written, we take the opportunity to create the release cache
directory.