feat: source-port ls command
This was something that seemed quite simple, but ended up needing quite a lot of code to support.
The 3 or so commits preceding this one are also features and changes required to support this new
command.

The main reason this became complicated was because I discovered that unauthenticated use of
Github's API has quite stringent limits. In retrospect it isn't really surprising, but it didn't
occur to me in advance. If you go over the fairly small limits, your IP will be blocked for a while.
I'm not sure what authenticated options may be appropriate, without asking users to sign up for a
Github account to use TDL.

The solution was to implement a cache for the release information, which is what we're querying
Github for. If the user runs a `source-port ls` command, each of the source port repositories are
queried for the latest release that's available. These results will be cached, with an expiry time
of 24 hours.

A FakeUtc struct was created in the test helpers crate so that the tests related to caching aren't
dependent on a particular time. This struct can be used with conditional compilation to return a
specific date and time when `Utc::now` is called.

A release repository trait was defined that has a method on it for getting the latest source port
release for a given repository. There's obviously an implementation of this trait that calls the
Github API, but there's also a 'fake' one that enabled unit testing. Using the fake repository and
saved responses from the Github API, it was possible to do unit testing for the caching
functionality. These responses have been committed as test data. They'll obviously go out of date,
but hopefully the Github releases shouldn't change that much.

There's an integration test that would actually run the command and that would hit Github, but this
has to be explicitly enabled using a Cargo feature, otherwise I would hit the API all the time and
probably end up getting banned.