refactor: specific source port and storage errors
Functions related to source port and storage now use their own error types.

This came about because I wanted to do something with a specific kind of error, namely when you're
trying to retrieve the latest release for a particular source port, but the owner of that repository
hasn't marked any as latest. This introduced a particular error enum for the source port, and in
order to be easily compatible with the use of the `?` operator, the `thiserror` crate was used.

I ended up refactoring the ObjectRepository struct to also use specific errors, since it was being
called from within the source port functions, so it didn't make sense for that to be using a
`Report` type for the error.