feat: handle case for repos with no latest release
When a source port repository has no release marked as latest, the behaviour is to return a specific
error to that effect. Even if the repo doesn't have a latest release, we still cache an entry for
it, because we wouldn't want to repeatedly hit the Github API. We just mark the version as
"no_latest_release".

When the cache is being queried for an entry, we check if that entry has its version set to
"no_latest_release", and if it does, we return an error.