feat: finalise `source-port ls` command
Finally, after all the refactoring and deliberation over the implementation of this feature, the
command handler is updated to present the information and there's an integration test that will hit
the Github API to make sure all the latest versions are returned (and source ports that don't have a
latest version will be indicated).

The caching part is all unit tested, so I'm confident that should be covered just by unit testing.