ci: run source port ls integration tests
These can run on every PR, so I'll be forced to update the tests when new versions of a source port
are released. This will make sure that the application is always returning the correct versions, and
I can respond if the source port owners decide to use weird version numbers or change asset names
and such things.