feat: `source-port install` command
Installs the latest version of the specified source port, by downloading the latest build from
Github and extracting it to `TDL_DOOM_HOME\<source-port>-<version>`. Example: `source-port install
Crispy`.

This command is only supported on Windows. I may extend it later to support macOS, but I don't think
it makes sense for Linux, where people would generally be using package managers to install builds
of Doom, or at least that's my understanding. Users on these operating systems would use the
`source-port add` command after installing the source port by some other means.

Given this command works with and extracts zip/tar archives, it was more amenable to integration
testing rather than unit testing. Unit testing would have necessitated lots of boilerplate that
seemed overkill.

The integration test runs the command and checks for the existence of the extracted exe, then runs
the `source-port ls` command to verify that TDL now considers it to be installed. This was actually
good because the `source-port ls` command had to be extended to indicate whether the source port had
been installed, so I ended up with more functionality.

It was important that these tests didn't hit the Github API, so the test sets up dummy cache
entries that are returned when the command queries for the latest release. This test data shouldn't
require any more maintenance than the other tests that are using Github responses.